# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'lciii-scaffolding'","description":"'LCII Initial Scaffolding'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8081","description":"'LCII Initial Scaffolding'"}],"paths":{"/loteria/apuesta":{"post":{"tags":["apuesta-controller"],"operationId":"createApuesta","parameters":[{"name":"dto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ApuestaDTORequest"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApuestaDTOResponse"}}}}}}},"/{fecha}":{"get":{"tags":["ping-controller"],"operationId":"ObtenerSorteoPorFecha","parameters":[{"name":"fecha","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SorteoPorFecha"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/loteria/total/{id_sorteo}":{"get":{"tags":["apuesta-controller"],"operationId":"getInformacionApuestasGanadas","parameters":[{"name":"id_sorteo","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InformacionApuestasGanadasDto"}}}}}}},"/loteria/sorteo/{id_sorteo}":{"get":{"tags":["apuesta-controller"],"operationId":"getInformacionSorteo","parameters":[{"name":"id_sorteo","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InformacionSorteo"}}}}}}}},"components":{"schemas":{"ApuestaDTORequest":{"type":"object","properties":{"fecha_sorteo":{"type":"string"},"id_cliente":{"type":"string"},"numero":{"type":"string"},"montoApostado":{"type":"integer","format":"int32"}}},"ApuestaDTOResponse":{"type":"object","properties":{"id_sorteo":{"type":"integer","format":"int32"},"fecha_sorteo":{"type":"string"},"id_cliente":{"type":"string"},"numero":{"type":"string"},"resultado":{"type":"string"}}},"SorteoPorFecha":{"type":"object","properties":{"numeroSorteo":{"type":"integer","format":"int32"},"fecha":{"type":"string"},"dineroTotalAcumulado":{"type":"integer","format":"int32"},"numerosSorteados":{"type":"array","items":{"type":"array","items":{"type":"integer","format":"int32"}}}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"InformacionApuestasGanadasDto":{"type":"object","properties":{"id_sorteo":{"type":"integer","format":"int32"},"fecha_sorteo":{"type":"string"},"totalDeApuestas":{"type":"integer","format":"int32"},"totalPagado":{"type":"integer","format":"int32"},"totalEnReserva":{"type":"integer","format":"int32"}}},"ApuestoInfoSorteoDto":{"type":"object","properties":{"id_cliente":{"type":"string"},"numero":{"type":"string"},"resultado":{"type":"string"},"montoApostado":{"type":"integer","format":"int32"},"premio":{"type":"integer","format":"int32"}}},"InformacionSorteo":{"type":"object","properties":{"id_sorteo":{"type":"integer","format":"int32"},"fecha_sorteo":{"type":"string"},"apuestoInfoSorteoDtos":{"type":"array","items":{"$ref":"#/components/schemas/ApuestoInfoSorteoDto"}},"totalEnReserva":{"type":"integer","format":"int32"}}}}}}